pyproject.toml
[project]
dependencies = [
  "livekit-agents[openai]~=1.0",
  "livekit-plugins-noise-cancellation~=0.2",
  "openai>=1.95.1",
  "pylint>=3.3.7",
  "python-dotenv>=1.1.1",
]
description = "Add your description here"
name = "livekit-agent"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"
